Java2DTextRenderer: Allow to override the font, which is choosen as fallback font, if some character is not in the current font.
The "Default" font is plattform dependent. Some Characters are not
contained in the default Font on each Plattform. I.e. ⚓ is not contained
in the default font on OS X, but its contained on Linux. For such special
characters it must be possible to override the way the fallback font is
choosen.